pyproject.toml
[project]
name = "simulation_guide"
version = "0.1.0"
description = "A guide to the simulation"
authors = [
    {name = "Abdullah Abrahams", email = "turnono@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.11"
requests = "^2.31.0"
google-adk = "^0.1.0"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.42.1"}
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"
google-cloud-firestore = "^2.20.2"

[tool.poetry.scripts]
simulation_guide = "server.serve:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

 